Initial load for c# DLL to call Report Writer
This code is designed to replace the existing legacy DLLs for Report Writer.  This code is a wrapper for the pdfrw64.exe program.  It allows for easily starting/stopping the server version of Report Writer as well as making calls to generate PDFs and retrieve a byte array (if desired) back for use in web pages, database storage, etc.